Se agrega filtros de búsqueda para Actividades y paginación de la misma

+++ SMGPA/SMGPA/Controllers/EntitiesController.cs
@@ -15,7 +15,7 @@ namespace SMGPA.Controllers
         private SMGPAContext db = new SMGPAContext();
 
         // GET: Entities
-        public ActionResult Index()
+        public ActionResult Index(string sortOrder)
         {
             return View(db.Entity.ToList());
         }
@@ -58,7 +58,73 @@ namespace SMGPA.Controllers
 
             return View(entities);
         }
+        // GET: Entities/CreateFaculty
+        public ActionResult CreateFaculty()
+        {
+            Faculty facultad = new Faculty();
+            facultad.Carreras = new List<Career>();
+            TempData["Facultad"] = facultad;
+            List<Career> Carreras = db.Career.ToList();
+            ViewBag.idCareer = new SelectList(Carreras.Where(c => c.idFaculty == null), "idCareer", "Nombre");
+            return View(facultad);
+        }
 
+        // POST: Entities/CreateFaculty
+        // Para protegerse de ataques de publicación excesiva, habilite las propiedades específicas a las que desea enlazarse. Para obtener 
+        // más información vea http://go.microsoft.com/fwlink/?LinkId=317598.
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult CreateFaculty([Bind(Include = "idEntities,Nombre,Descripcion,Activo")] Faculty faculty)
+        {
+                List<Career> _Carreras = new List<Career>();
+                List<Career> Carreras = db.Career.ToList();
+                if (ModelState.IsValid)
+                {
+                    Guid idEntities = Guid.NewGuid();
+                    faculty.idEntities = idEntities;
+                    List<Career> Careers = (List<Career>)TempData["Carreras"];
+                    //Si viene sin carreras retorno mensaje indicando que no es posible
+                    if (Careers == null)
+                    {
+                        ViewBag.Errores = "No puede crear una Facultad sin Carreras asociadas";
+                        ViewBag.idCareer = new SelectList(Carreras.Where(c => c.idFaculty == null), "idCareer", "Nombre");
+                        return View(faculty);
+                    }
+                    foreach (Career c in Careers)
+                    {
+                        Career car = db.Career.Find(c.idCareer);
+                        car.idFaculty = idEntities;
+                        List<Functionary> funcionarios = db.Functionary.Where(f => f.idCareer == c.idCareer).ToList();
+                        if (funcionarios != null)
+                        {
+                            foreach (Functionary f in funcionarios)
+                            {
+                                faculty.Involucrados.Add(f);
+                            }
+                        }
+                    }
+
+                    db.Faculty.Add(faculty);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                ViewBag.idCareer = new SelectList(Carreras.Where(c => c.idFaculty == null), "idCareer", "Nombre");
+                return View(faculty);
+        }
+        [HttpPost]
+        public  async Task<JsonResult> AddCareer(Guid id)
+        {
+            if (id == null)
+            {
+                return Json(new { sucess = false }, JsonRequestBehavior.AllowGet);
+            }
+            Faculty facultad = (Faculty)TempData["Facultad"];
+            Career carrera = await db.Career.FindAsync(id);
+            facultad.Carreras.Add(carrera); 
+            TempData["Carreras"] = facultad.Carreras;
+            TempData["Faculta0.d"] = facultad;
+            return Json(new { sucess = true , nombre = carrera.Nombre, idcareer = carrera.idCareer}, JsonRequestBehavior.AllowGet);
+        }
         // GET: Entities/Edit/5
         public ActionResult Edit(Guid? id)
         {
@@ -110,9 +176,25 @@ namespace SMGPA.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(Guid id)
         {
-            Entities entities = db.Entity.Find(id);
-            db.Entity.Remove(entities);
-            db.SaveChanges();
+            Faculty facultad = db.Faculty.Find(id);
+            if(facultad != null)
+            {
+                if(facultad.Carreras != null)
+                {
+                    foreach(Career c in facultad.Carreras)
+                    {
+                        c.idFaculty = null;
+                    }
+                }
+                db.Faculty.Remove(facultad);
+                db.SaveChanges();
+            }
+            if(facultad == null)
+            {
+                Entities entities = db.Entity.Find(id);
+                db.Entity.Remove(entities);
+                db.SaveChanges();
+            }
             return RedirectToAction("Index");
         }
         [HttpGet]
